HSEARCH-3118 Move low-level Lucene code to a separate package
We will have to have dependencies from writers to readers at some point
(for NRT), so we really shouldn't put the reader-related classes in the
search package.
