+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/lowlevel/directory/impl/MMapDirectoryProvider.java
@@ -4,14 +4,13 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.lucene.impl;
+package org.hibernate.search.backend.lucene.lowlevel.directory.impl;
 
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-import org.hibernate.search.backend.lucene.index.impl.DirectoryProvider;
 import org.hibernate.search.backend.lucene.logging.impl.Log;
 import org.hibernate.search.util.common.reporting.EventContext;
 import org.hibernate.search.util.common.logging.impl.LoggerFactory;
@@ -19,14 +18,14 @@ import org.hibernate.search.util.common.logging.impl.LoggerFactory;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.MMapDirectory;
 
-class MMapDirectoryProvider implements DirectoryProvider {
+public class MMapDirectoryProvider implements DirectoryProvider {
 	private static final Log log = LoggerFactory.make( Log.class, MethodHandles.lookup() );
 
 	private final EventContext backendContext;
 
 	private final Path rootDirectory;
 
-	MMapDirectoryProvider(EventContext backendContext, Path rootDirectory) {
+	public MMapDirectoryProvider(EventContext backendContext, Path rootDirectory) {
 		this.backendContext = backendContext;
 		this.rootDirectory = rootDirectory;
 	}
